Added handler methods for no fdData content
When a DeviceCloud push notification comes through and there is no fdData content, this can mean that the payload was over the limit (120KB). We still want to be able to handle this notification and have a "backup plan" of sorts (maybe fetch the data in a background process). This happens a lot for data notifications when an image is too large to send encoded.